+++ src/stores/api/inst-data.ts
@@ -3,16 +3,16 @@ import { storedActiveSelection } from '../states';
 import { LOADED, LOADING } from "../../_interfaces";
 
 export const loadInstData = async () => {
-    storedActiveSelection.update(value => {
-        value.loadedData.inst = LOADING;
-        return value;
+    storedActiveSelection.update(store => {
+        store.loadedData.inst = LOADING;
+        return store;
     });
     // @ts-ignore TS2339
     const data = await fetch(`/data/inst.json?cb=${window.cacheBuster}`).then(res => res.json());
 
-    storedActiveSelection.update(value => {
-        value.loadedData.inst = LOADED;
-        return value;
+    storedActiveSelection.update(store => {
+        store.loadedData.inst = LOADED;
+        return store;
     });
 
     // { "100247": "/100/100247 Tragschnabelwagen (36MB).pdf", ... }